backend/requirements.txt
Flask
requests
flask-cors
python-dotenv
psycopg2-binary
SQLAlchemy